refactored TestAssertionHelper... and merged with AssertionHelpers
